+++ python/lsst/ts/cRIOpy/VMS/__init__.py
@@ -2,6 +2,7 @@ from .Bars import StatusBar, ToolBar
 from .BoxChartWidget import BoxChartWidget
 from .Cache import Cache
 from .ChartView import ChartView
+from .DisplacementWidget import DisplacementWidget
 from .PSDWidget import PSDWidget
 from .RawAccelerationWidget import RawAccelerationWidget
 from .TimeBoxChart import TimeBoxChart
